requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.3.0
streamlit>=1.2.0
tqdm>=4.62.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
xgboost>=1.5.0
kaggle>=1.5.0